requirements.txt
SpeechRecognition
pyttsx3
gTTS
playsound
pywhatkit
wikipedia-api
platformdirs
datetime
